add handler for android platform specific AndroidViewEvent
This fixes broken foreground service handling on android.
The previous behavior with android was broken because when swiped off tray and re-launched, the application was still running and had not exited properly.

On this branch, this fixes this behavior so that when katzen is launched the notification appears, and when the view is swiped off tray, immediately stops.

The issue was that in gio, when an app.DestroyEvent is received, while a foreground service is running, it is too late to stop the foreground service before
android has decided it shouldn't stop the main activity after all.

The workaround is that in gio, when an android view onDestroy is received by GioView, gio sends an AndroidViewEvent.View = 0
when this occurs, we should stop reading Events, and halt the foreground service, otherwise
the foreground service is still running and android does not halt the main activity at onDestroy.

If this isn't done, immediately re-launching the application tries to open a new view with the existing activity,
which causes a ANR as the view does not receive input. Otherwise, the ANR eventually kills the process and re-launching works again.

I tested this and confirmed that on android-31:
I can start/stop katzen, the foreground service notification is displayed, and when the view is swiped off tray, immediately is removed.
re-starting katzen presents the login screen as expected.
